Moved async data-fetching into new FeaturedList component with CardListSkeleton fallback
