+++ src/app/page.tsx
@@ -1,30 +1,12 @@
-import { CardList, RandomPokemonButton } from '@/components/pokemon';
+import { RandomPokemonButton } from '@/components/pokemon';
 import { SearchBar } from '@/components/ui/';
-import {
-  createPokemonPromises,
-  getAllPokemonNames,
-  getRandomPokemonNames,
-} from '@/lib/data/rest-api/pokemon';
 import { navigateToSearchedPokemon } from '@/lib/actions/serverActions';
+import { Suspense } from 'react';
+import { CardListSkeleton } from '@/components/skeletons/card-list-skeleton';
+import { PokemonSearchBar } from '@/components/pokemon/pokemon-search-bar';
+import { FeaturedList } from '@/components/pokemon/featured-list';
 
-export default async function Home() {
-  // Get list of results for SearchBar
-  let pokemonList: string[] | undefined = undefined;
-  const pokemonListResult = await getAllPokemonNames();
-  if (pokemonListResult.success) {
-    pokemonList = pokemonListResult.data;
-  }
-
-  // Get PokemonPromises for FeaturedList, with fallback list
-  let featuredPokemon = ['pikachu', 'venasaur', 'charizard', 'blastoise'];
-
-  const randomResult = await getRandomPokemonNames(4);
-  if (randomResult.success) {
-    featuredPokemon = randomResult.data;
-  }
-
-  const featuredPromises = createPokemonPromises(featuredPokemon);
-
+export default function Home() {
   return (
     <>
       <section className="content-grid full-width items-center gap-4 [background-image:linear-gradient(-10deg,_#C97FE4,_#AECDF6)]">
@@ -52,7 +34,9 @@ export default async function Home() {
       </section>
       <section className="content-grid full-width [background-image:linear-gradient(-10deg,_#f5e6fb,_#eef5fd)] pb-12">
         <h2 className="text-bold text-center text-4xl p-8">Featured Pokémon</h2>
-        <CardList pokemonPromises={featuredPromises} />
+        <Suspense fallback={<CardListSkeleton numCards={4} />}>
+          <FeaturedList number={4} />
+        </Suspense>
       </section>
     </>
   );
