+++ src/components/skeletons/pokemon-card-skeleton.tsx
@@ -1,5 +1,7 @@
 import { pokemonCardStyle } from '../pokemon';
 
 export function PokemonCardSkeleton() {
-  return <div className={`${pokemonCardStyle}`}></div>;
+  return (
+    <div className={`${pokemonCardStyle} bg-neutral-100 animate-pulse`}></div>
+  );
 }
